requirements.txt
Django==1.9.4
beautifulsoup4==4.4.1
djangorestframework==3.3.2
psycopg2==2.6.1
